requirements.txt
google-cloud-bigquery-datatransfer
google-cloud-bigquery
google-cloud-pubsub
Flask
gunicorn
werkzeug
flask-compress
